last changes update min credits

+++ app/Http/Controllers/SubjectController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\SubjectRequest;
+use App\Models\Student;
 use App\Models\Subject;
 
 class SubjectController extends Controller
@@ -73,8 +74,16 @@ class SubjectController extends Controller
      */
     public function destroy(Subject $subject)
     {
+        $hasStudents = Student::whereHas('subjects', function ($query) use ($subject) {
+            $query->where('subject_id', $subject->id);
+        })->exists();
+
+        if ($hasStudents) {
+            return response()->json(['error' => 'This subject has associated students and cannot be eliminated.'], 422);
+        }
+
         $subject->delete();
-        return redirect()->route('subject.index');
+        return response()->json(['message' => 'Subject successfully deleted.']);
     }
 
     public function getSubjectsWithProfessors()
